modify
git-svn-id: file:///var/www/svn/mxtce-sw@238 52860be4-34ab-47e5-b986-9ec2afffd7f9

+++ src/api/reply_encoder.cc
@@ -366,10 +366,10 @@ void Apireplymsg_encoder::encode_path(TPath* path_info, Encode_Pri_Type* pri_typ
 			//sw_cap_descriptors=(*it)->GetTheISCD();
 
 			switchingType = (*iscdVar)->switchingType;
-			cout<<"switchingtype="<<switchingType<<endl;
+			cout<<"switchingtype="<<(int)switchingType<<endl;
 
 			encodingType = (*iscdVar)->encodingType;
-			cout<<"encodingtype="<<encodingType<<endl;
+			cout<<"encodingtype="<<(int)encodingType<<endl;
 
 			//always encode switch capability type first
 			switch (switchingType)
